Parse interface speed as float (#24775)

+++ pkg/collector/corechecks/network-devices/cisco-sdwan/payload/interfaces.go
@@ -18,7 +18,7 @@ type CiscoInterface interface {
 	// Index returns the interface index
 	Index() (int, error)
 	// GetSpeedMbps returns the interface speed in megabits per second
-	GetSpeedMbps() int
+	GetSpeedMbps() float64
 	// OperStatus returns the interface oper status
 	OperStatus() devicemetadata.IfOperStatus
 	// AdminStatus returns the interface admin status
